+++ tests/test_config.py
@@ -1,4 +1,7 @@
 from collections import OrderedDict
+from pathlib import Path
+from typing import Iterator
+import os
 
 import pytest
 
@@ -127,3 +130,331 @@ def test_tool_no_explicit_tool(tmpdir):
     c = Config(str(p))
 
     assert c.get_tool("testtool") == "testtool"
+
+def test_include(tmp_path: Path) -> None:
+    config_yaml = tmp_path / "configs" / "config.yaml"
+    config_yaml.parent.mkdir(parents=True, exist_ok=True)
+    config_yaml.write_text("""---
+target:
+  main:
+    drivers: {}
+includes:
+  - !include first.yaml
+  - !include second.yaml
+""")
+    first_yaml = tmp_path / "configs" / "first.yaml"
+    first_yaml.write_text("""
+first:
+  - a
+  - b
+  - c
+""")
+
+    second_yaml = tmp_path / "configs" / "second.yaml"
+    second_yaml.write_text("""
+second:
+  foo: bar
+""")
+
+    c = Config(str(config_yaml.resolve()))
+    assert 'b' in c.data['first']
+    assert c.data['second']['foo'] == 'bar'
+
+@pytest.fixture(scope="function")
+def include_env() -> Iterator[None]:
+    os.environ['LG_FIRST'] = 'first.yaml'
+    os.environ['LG_SECOND'] = 'second.yaml'
+    os.environ['LG_THIRD'] = 'third'
+    os.environ['LG_FOO'] = 'bar'
+    yield
+    del os.environ['LG_FOO']
+    del os.environ['LG_THIRD']
+    del os.environ['LG_SECOND']
+    del os.environ['LG_FIRST']
+
+def test_include_inline(tmp_path: Path, include_env: None) -> None:
+    del include_env  # unused
+
+    config_yaml = tmp_path / "configs" / "config.yaml"
+    config_yaml.parent.mkdir(parents=True, exist_ok=True)
+    config_yaml.write_text("""---
+target:
+  main:
+    drivers: {}
+  subnode: !include first.yaml
+  another: !include second.yaml
+""")
+    first_yaml = tmp_path / "configs" / "first.yaml"
+    first_yaml.write_text("""---
+first:
+  - a
+  - b
+  - c
+""")
+
+    second_yaml = tmp_path / "configs" / "second.yaml"
+    second_yaml.write_text("""---
+foo: !template ${LG_FOO}
+""")
+
+    c = Config(str(config_yaml.resolve()))
+    assert 'a' in c.data['target']['subnode']['first']
+    assert c.data['target']['another']['foo'] == 'bar'
+
+def test_include_inline_var(tmp_path: Path, include_env: None) -> None:
+    del include_env  # unused
+
+    config_yaml = tmp_path / "configs" / "config.yaml"
+    config_yaml.parent.mkdir(parents=True, exist_ok=True)
+    config_yaml.write_text("""---
+target:
+  main:
+    drivers: {}
+  subnode: !include ${LG_FIRST}
+  another: !include $LG_SECOND
+  yetanother: !include ${LG_THIRD}.yaml
+""")
+    first_yaml = tmp_path / "configs" / "first.yaml"
+    first_yaml.write_text("""---
+first:
+  - a
+  - b
+  - c
+""")
+
+    second_yaml = tmp_path / "configs" / "second.yaml"
+    second_yaml.write_text("""---
+foo: !template ${LG_FOO}
+""")
+
+    third_yaml = tmp_path / "configs" / "third.yaml"
+    third_yaml.write_text("""---
+hello: world
+""")
+
+    c = Config(str(config_yaml.resolve()))
+    assert 'a' in c.data['target']['subnode']['first']
+    assert c.data['target']['another']['foo'] == 'bar'
+    assert c.data['target']['yetanother']['hello'] == 'world'
+
+def test_include_inline_var_doesnotexist(tmp_path: Path, include_env: None) -> None:
+    del include_env  # unused
+
+    config_yaml = tmp_path / "configs" / "config.yaml"
+    config_yaml.parent.mkdir(parents=True, exist_ok=True)
+    config_yaml.write_text("""---
+target:
+  main:
+    drivers: {}
+  subnode: !include ${LG_DOESNOTEXIST}
+  another: !include ${LG_FIRST}
+""")
+    first_yaml = tmp_path / "configs" / "first.yaml"
+    first_yaml.write_text("""---
+first:
+  - a
+  - b
+  - c
+""")
+
+    with pytest.raises(InvalidConfigError) as excinfo:
+        Config(str(config_yaml.resolve()))
+    assert 'Could not resolve key' in excinfo.value.msg
+
+def test_include_var(tmp_path: Path, include_env: None) -> None:
+    del include_env  # unused
+
+    config_yaml = tmp_path / "configs" / "config.yaml"
+    config_yaml.parent.mkdir(parents=True, exist_ok=True)
+    config_yaml.write_text("""---
+target:
+  main:
+    drivers: {}
+includes:
+  - !include ${LG_FIRST}
+""")
+    first_yaml = tmp_path / "configs" / "first.yaml"
+    first_yaml.write_text("""
+first:
+  - a
+  - b
+  - c
+""")
+
+    c = Config(str(config_yaml.resolve()))
+    assert 'b' in c.data['first']
+
+def test_include_template(tmp_path: Path, include_env: None) -> None:
+    del include_env  # unused
+
+    config_yaml = tmp_path / "configs" / "config.yaml"
+    config_yaml.parent.mkdir(parents=True, exist_ok=True)
+    config_yaml.write_text("""---
+target:
+  main:
+    drivers: {}
+includes:
+  - !include first.yaml
+""")
+    first_yaml = tmp_path / "configs" / "first.yaml"
+    first_yaml.write_text("""
+first:
+  - foo
+  - !template ${LG_FOO}
+  - baz
+""")
+
+    c = Config(str(config_yaml.resolve()))
+    assert 'bar' in c.data['first']
+
+def test_include_template_bad_placeholder(tmp_path: Path, include_env: None) -> None:
+    del include_env  # unused
+
+    config_yaml = tmp_path / "configs" / "config.yaml"
+    config_yaml.parent.mkdir(parents=True, exist_ok=True)
+    config_yaml.write_text("""---
+target:
+  main:
+    drivers: {}
+includes:
+  - !include first.yaml
+""")
+    first_yaml = tmp_path / "configs" / "first.yaml"
+    first_yaml.write_text("""
+first:
+  - foo
+  - !template $
+  - baz
+""")
+
+    with pytest.raises(InvalidConfigError) as excinfo:
+        Config(str(config_yaml.resolve()))
+    assert "is invalid" in excinfo.value.msg
+    assert "template string" in excinfo.value.msg
+
+def test_include_template_bad_key(tmp_path: Path, include_env: None) -> None:
+    del include_env  # unused
+
+    config_yaml = tmp_path / "configs" / "config.yaml"
+    config_yaml.parent.mkdir(parents=True, exist_ok=True)
+    config_yaml.write_text("""---
+target:
+  main:
+    drivers: {}
+includes:
+  - !include first.yaml
+""")
+    first_yaml = tmp_path / "configs" / "first.yaml"
+    first_yaml.write_text("""
+first:
+  - foo
+  - !template ${LG_DOESNOTEXIST}
+  - baz
+""")
+
+    with pytest.raises(InvalidConfigError) as excinfo:
+        Config(str(config_yaml.resolve()))
+    assert 'refers to unknown variable' in excinfo.value.msg
+
+def test_include_template_var(tmp_path: Path, include_env: None) -> None:
+    del include_env  # unused
+
+    config_yaml = tmp_path / "configs" / "config.yaml"
+    config_yaml.parent.mkdir(parents=True, exist_ok=True)
+    config_yaml.write_text("""---
+target:
+  main:
+    drivers: {}
+includes:
+  - !include ${LG_FIRST}
+""")
+    first_yaml = tmp_path / "configs" / "first.yaml"
+    first_yaml.write_text("""
+first:
+  - foo
+  - !template ${LG_FOO}
+  - baz
+""")
+
+    c = Config(str(config_yaml.resolve()))
+    assert 'bar' in c.data['first']
+
+def test_include_template_var_bad_placeholder(tmp_path: Path, include_env: None) -> None:
+    del include_env  # unused
+
+    config_yaml = tmp_path / "configs" / "config.yaml"
+    config_yaml.parent.mkdir(parents=True, exist_ok=True)
+    config_yaml.write_text("""---
+target:
+  main:
+    drivers: {}
+includes:
+  - !include ${LG_FIRST}
+""")
+    first_yaml = tmp_path / "configs" / "first.yaml"
+    first_yaml.write_text("""
+first:
+  - foo
+  - !template $
+  - baz
+""")
+
+    with pytest.raises(InvalidConfigError) as excinfo:
+        Config(str(config_yaml.resolve()))
+    assert "is invalid" in excinfo.value.msg
+    assert "template string" in excinfo.value.msg
+
+def test_include_template_var_bad_key(tmp_path: Path, include_env: None) -> None:
+    del include_env  # unused
+
+    config_yaml = tmp_path / "configs" / "config.yaml"
+    config_yaml.parent.mkdir(parents=True, exist_ok=True)
+    config_yaml.write_text("""---
+target:
+  main:
+    drivers: {}
+includes:
+  - !include ${LG_FIRST}
+""")
+    first_yaml = tmp_path / "configs" / "first.yaml"
+    first_yaml.write_text("""
+first:
+  - foo
+  - !template ${LG_DOESNOTEXIST}
+  - baz
+""")
+
+    with pytest.raises(InvalidConfigError) as excinfo:
+        Config(str(config_yaml.resolve()))
+    assert "refers to unknown variable" in excinfo.value.msg
+
+def test_include_template_var_doesnotexist(tmp_path: Path, include_env: None) -> None:
+    del include_env  # unused
+
+    config_yaml = tmp_path / "configs" / "config.yaml"
+    config_yaml.parent.mkdir(parents=True, exist_ok=True)
+    config_yaml.write_text("""---
+target:
+  main:
+    drivers: {}
+includes:
+  - !include first.yaml
+  - !include ${LG_DOESNOTEXIST}second.yaml
+""")
+    first_yaml = tmp_path / "configs" / "first.yaml"
+    first_yaml.write_text("""
+first:
+  - foo
+  - bar
+  - baz
+""")
+    second_yaml = tmp_path / "configs" / "second.yaml"
+    second_yaml.write_text("""
+second:
+  foo: bar
+""")
+
+
+    with pytest.raises(InvalidConfigError) as excinfo:
+        Config(str(config_yaml.resolve()))
+    assert "Could not resolve key" in excinfo.value.msg